Add configuration values for capacity, transfer rate, energy consumption per tick per ingredient, recipe duration of the auto crafter

+++ src/main/java/me/jddev0/ep/block/entity/EnergizerBlockEntity.java
@@ -227,7 +227,7 @@ public class EnergizerBlockEntity extends BlockEntity implements ExtendedScreenH
                 blockEntity.energyConsumptionLeft = energyConsumption;
 
             //TODO improve (alternate values +/- 1 per x recipes instead of changing last energy consumption tick)
-            long energyConsumptionPerTick = (int)Math.ceil((float)energyConsumption / blockEntity.maxProgress);
+            long energyConsumptionPerTick = (long)Math.ceil((double)energyConsumption / blockEntity.maxProgress);
             if(blockEntity.progress == blockEntity.maxProgress - 1)
                 energyConsumptionPerTick = blockEntity.energyConsumptionLeft;
